+++ async-race/src/app/core/components/footer/_footer.scss
