+++ async-race/src/app/core/components/footer/footer.ts
