+++ async-race/src/app/features/components/car-cell/_car-cell.scss
