+++ async-race/src/app/features/components/winner-cell/_winner-cell.scss
