+++ async-race/src/app/features/components/winner-cell/winner-cell.ts
