+++ async-race/src/app/features/components/winners/_winners.scss
