+++ async-race/src/app/shared/components/car-input/_car-input.scss
