+++ async-race/src/app/shared/components/pagination/_pagination.scss
